# Update ghost particles without rebuilding them