This is a work in progress ([WIP])
This PR aims at allowing for a function that updates the ghost particles without rebuilding them from scratch. Many numerical methods don't require that the particle interact with the mesh at every time step (e.g. MD based methods or DEM). This PR aims at addressing #11093 .

- [X] Create an update_ghost_particles function and move the necessary containers to particle_handler class member. In the update_ghost_particles remove the loop to generate the ghost_particles_by_domain
- [x] Add a unit test for ghost updates
- [ ] Modify send_rcv_particles (or write a new function alltogether) to enable updating existing particles without recreating them from scratch. This part i will do with the advices @peterrum (thanks again for helping me with this :).

The first modification already implemented allowed us a speed up of 100% in our DEM code on 2 cores (from 280s to 127s) and a 30% (103s to 72s) at 16 core. So this is already extremely significant. I think it's the last optimization we need in our case. This will also make the particle_handler a very very flexible class!