@peterrum 2020-10-31
> The second issue is more profound. At every call, exchange_ghost_particles() clears the entire ghost_particles structure and rebuilds it from scratch, assuming that ghost particles may have changed cells and everything. The underlying assumption is that we are calling sort_particles_into_subdomain_and_cells at every time step. Regretfully, this is not what we are doing.

@blaisb Could you quickly explain me the reason for why you don't have to call `sort_particles_into_subdomain_and_cells` in the case of DEM every time step. Maybe this is a stupid question and I am not aware about the differences of DEM and PIC.