@blaisb 2020-10-31
> > The second issue is more profound. At every call, exchange_ghost_particles() clears the entire ghost_particles structure and rebuilds it from scratch, assuming that ghost particles may have changed cells and everything. The underlying assumption is that we are calling sort_particles_into_subdomain_and_cells at every time step. Regretfully, this is not what we are doing.
> 
> @blaisb Could you quickly explain me the reason for why you don't have to call `sort_particles_into_subdomain_and_cells` in the case of DEM every time step. Maybe this is a stupid question and I am not aware about the differences of DEM and PIC.

In DEM  the particle interact between themselves every timestep (we do particle-particle force calculation through either contact models or non-contact models (e.g. an LJ potential)). However, we want the particles to interact with the mesh every **n** iterations, where **n** is a number between 1 and 1000. Consequently, when we calculate particle-particle interaction, we don't need to have the particles exactly sorted into the subdomain (we don't need their reference location). The only thing we need is their properties (velocity, diameter) and their physical location. This is sufficient to do the particle-particle interaction. Thus, we need a mechanism to update the properties and the physical location of the ghost particles.

Right now we are using exchange_ghost to achieve this. But exchange ghost has a very large overhead. Right now, a single core DEM benchmark simulation in Lethe runs in 120s on one core, but runs in 287s on two cores because of the overhead of exchange_ghost_particles. The present PR in it's current state brings the two-core cost to 120s, removing 160s of overhead just because of the loop that is removed when calling update_ghosts every iteration and exchange ghost every **n** iterations.

I know this is usage is not what was originally intended for particles. However, I think providing this update mechanism does not break the code for anyone. It provides a much more flexible implementation that allow a code to do particle-particle stuff at one time scale and particle-mesh stuff at another time scale. Our ultimate goal is to be able to build the first families of DEM scheme where DEM can be coupled with a CFD solver without requiring that we couple two existing code (e.g. OpenFOAM and LIGGGHTS, Fluent and EDEM, etc.). We think this will provide at least a 10x speed-up over the current state of the art just by using existing deal.II facilities and what we added in Lethe.

I hope that clearly explains it :)!.