@peterrum 2020-10-31
Thanks for the explanation @blaisb! If I see it correctly you make the simplified assumption that no new ghost-cell particles are created during the n time steps, which is probably acceptable in your case. We can discuss this later offline.