@blaisb 2020-11-03
This last push finishes task 2 of this PR.
I added a unit test which has a property and a position, which is changed and the ghosts are updated. It is simple, but it really does the job. I also applied @peterrum  comments and fixed indentation.