@blaisb 2020-11-03
> Thanks!

No problem. If it is good for you, I suggest we go as you suggested. We could merge this PR and then together work on the third (and last) item. :)!